+++ lua/ergoterm/commandline.lua
@@ -3,8 +3,8 @@ local lazy = require("ergoterm.lazy")
 
 ---@module "ergoterm.utils"
 local utils = lazy.require("ergoterm.utils")
----@module "ergoterm.text_decorators"
-local text_decorators = lazy.require("ergoterm.text_decorators")
+---@module "ergoterm.config"
+local config = lazy.require("ergoterm.config")
 
 local M = {}
 
@@ -209,7 +209,7 @@ M._all_options = {
   end,
 
   decorator = function(typed_decorator)
-    local decorators = vim.tbl_values(text_decorators.DECORATORS)
+    local decorators = vim.tbl_keys(config.get_text_decorators())
     if utils.str_is_empty(typed_decorator) then return decorators end
     return vim.tbl_filter(
       function(decorator) return decorator:match("^" .. typed_decorator .. "*") ~= nil end,